refactor(fedora): update and consolidate iso variables for Fedora 39
This commit updates the ISO URL and checksum variables for Fedora 39 to be more specific for different architectures (x86_64 and arm64). It also removes outdated templates for Fedora 36, which is no longer supported. The changes improve maintainability by reducing redundancy and ensuring the templates are up-to-date with the latest Fedora releases.