feat: update Packer templates and configurations for Fedora 40
This commit updates the Packer templates, scripts, and configurations to support Fedora 40. It includes the removal of Fedora 39-specific files and the addition of new files for Fedora 40, such as kickstart configurations, Packer templates, and variable files. The Makefile and README.md have also been updated to reflect the change from Fedora 39 to Fedora 40. This ensures compatibility with the latest Fedora release and maintains the ability to build Vagrant box images for both x86_64 and arm64 architectures.